requirements.txt
# Web framework
Flask>=2.3.0

# Core libraries
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.2
tqdm>=4.65.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.7.0

# Deep learning
torch>=2.1.0
torchvision>=0.16.0

# Similarity search
faiss-cpu>=1.7.4
